requirements.txt
pandas
mysql-connector-python
<<<<<<< HEAD
python-dotenv
=======
python-dotenv
pymysql
>>>>>>> d216e9f (Reinitialize Git in the correct directory)
